separate scripts from rulesets for specific application/reuse
