+++ src/configs/markdown.js
@@ -1,7 +1,7 @@
 import markdown from 'eslint-plugin-markdown'
 import tseslint from 'typescript-eslint'
 
-import { BUILD_IGNORES, MD_FILES } from './constants.js'
+import { BUILD_IGNORES, MD_FILES } from '../constants.js'
 
 // npx @eslint/config-inspector
 const configs = tseslint.config(
