+++ src/rulesets/index.js
@@ -1,0 +1,8 @@
+import jsOnly from './jsOnly.js'
+import jsOnlyScripts from './jsOnlyScripts.js'
+
+const ruleset = {
+  jsOnly,
+  jsOnlyScripts,
+}
+export default ruleset
